Adds files for testing interactive behaviour
not all functionality can be tested with pytest's unit tests (at least
not afaik). So an extra file is added that can be ran with ipython or in
the interactive window of (e.g.) VSCode or can be converted to a
Jupyter notebook. Test with printing a dataframe to the screen that
 contains phreeqpython solutions as column fails due to upstream
  bug. Pull request pending.